tests/wpt/tests/css/css-gaps/parsing/gap-decorations-rule-shorthand.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Gap Decorations: *-rule sets longhands</title>
<link rel="help" href="https://drafts.csswg.org/css-multicol/#propdef-column-rule">
<meta name="assert" content="column-rule supports the full grammar '<gap-rule-list> | <gap-auto-rule-list>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/shorthand-testcommon.js"></script>
</head>
<body>
<script>
// TODO(samomekarajr): Add `row-rule` to this test.
const rule_properties = {
    'column-rule': ['column-rule-width',
                    'column-rule-style',
                    'column-rule-color'],
};

for(rule_property in rule_properties) {
    const [width, style, color] = rule_properties[rule_property];
    // <gap-rule> = [<line-width> || <line-style> || <line-color>].
    test_shorthand_value(rule_property, '5px solid red', {
        [width]: '5px',
        [style]: 'solid',
        [color]: 'red'
    });

    test_shorthand_value(rule_property, 'double', {
        [width]: 'medium',
        [style]: 'double',
        [color]: 'currentcolor'
    });

    test_shorthand_value(rule_property, 'blue 10px', {
        [width]: '10px',
        [style]: 'none',
        [color]: 'blue'
    });

    // <gap-auto-repeat-rule> = repeat( auto , <gap-rule># ).
    test_shorthand_value(rule_property, 'repeat(auto, 5px solid green)', {
        [width]: 'repeat(auto, 5px)',
        [style]: 'repeat(auto, solid)',
        [color]: 'repeat(auto, green)'
    });

    test_shorthand_value(rule_property, 'repeat(auto, 5px solid yellow, 10px dotted blue)', {
        [width]: 'repeat(auto, 5px 10px)',
        [style]: 'repeat(auto, solid dotted)',
        [color]: 'repeat(auto, yellow blue)'
    });

    test_shorthand_value(rule_property, 'repeat(auto, blue 6px, 5px solid red)', {
        [width]: 'repeat(auto, 6px 5px)',
        [style]: 'repeat(auto, none solid)',
        [color]: 'repeat(auto, blue red)'
    });

    // <gap-repeat-rule> = repeat( <integer [1,∞]> , <gap-rule># ).
    test_shorthand_value(rule_property, 'repeat(4, 15px dotted pink)', {
        [width]: 'repeat(4, 15px)',
        [style]: 'repeat(4, dotted)',
        [color]: 'repeat(4, pink)'
    });
    test_shorthand_value(rule_property, 'repeat(1, 15px ridge yellow, 10px dotted blue, 15px double green)', {
        [width]: 'repeat(1, 15px 10px 15px)',
        [style]: 'repeat(1, ridge dotted double)',
        [color]: 'repeat(1, yellow blue green)'
    });
    test_shorthand_value(rule_property, 'repeat(3, lime 16px, dashed purple, 10px dotted)', {
        [width]: 'repeat(3, 16px medium 10px)',
        [style]: 'repeat(3, none dashed dotted)',
        [color]: 'repeat(3, lime purple currentcolor)'
    });

    // <gap-rule-list> = <gap-rule-or-repeat>#.
    // <gap-rule-or-repeat> = <gap-rule> | <gap-repeat-rule>.
    test_shorthand_value(rule_property, 'thin, dashed, hotpink', {
        [width]: 'thin medium medium',
        [style]: 'none dashed none',
        [color]: 'currentcolor currentcolor hotpink'
    });
    test_shorthand_value(rule_property, '5px double salmon, repeat(4, 5px ridge red)', {
        [width]: '5px repeat(4, 5px)',
        [style]: 'double repeat(4, ridge)',
        [color]: 'salmon repeat(4, red)'
    });
    test_shorthand_value(rule_property,
    'repeat(2, dashed gray, 10px blue dotted, 20px double), 5px solid red, repeat(4, blue 6px, 5px solid white)', {
        [width]: 'repeat(2, medium 10px 20px) 5px repeat(4, 6px 5px)',
        [style]: 'repeat(2, dashed dotted double) solid repeat(4, none solid)',
        [color]: 'repeat(2, gray blue currentcolor) red repeat(4, blue white)'
    });
    test_shorthand_value(rule_property, 'repeat(4, thick hidden skyblue), repeat(3, 5px solid red, 10px dotted)', {
        [width]: 'repeat(4, thick) repeat(3, 5px 10px)',
        [style]: 'repeat(4, hidden) repeat(3, solid dotted)',
        [color]: 'repeat(4, skyblue) repeat(3, red currentcolor)'
    });

    // <gap-auto-rule-list>   = <gap-rule-or-repeat>#? ,
    //                          <gap-auto-repeat-rule> ,
    //                          <gap-rule-or-repeat>#?.
    test_shorthand_value(rule_property,
    'repeat(auto, 10px solid red), medium dotted green, repeat(3, thick dashed blue, 15px double green)', {
        [width]: 'repeat(auto, 10px) medium repeat(3, thick 15px)',
        [style]: 'repeat(auto, solid) dotted repeat(3, dashed double)',
        [color]: 'repeat(auto, red) green repeat(3, blue green)'
    });

    test_shorthand_value(rule_property, 'ridge red, repeat(auto, 5px solid green), 10px dotted blue', {
        [width]: 'medium repeat(auto, 5px) 10px',
        [style]: 'ridge repeat(auto, solid) dotted',
        [color]: 'red repeat(auto, green) blue'
    });

    test_shorthand_value(rule_property,
    '10px dotted salmon, repeat(4, thin blue, hidden 5px purple), repeat(auto, 5px solid red, teal)', {
        [width]: '10px repeat(4, thin 5px) repeat(auto, 5px medium)',
        [style]: 'dotted repeat(4, none hidden) repeat(auto, solid none)',
        [color]: 'salmon repeat(4, blue purple) repeat(auto, red teal)'
    });
}
</script>
</body>
</html>
